Refactor initializeOptionsAndConfig
