+++ packages/cli/src/cmds/beacon/handler.ts
@@ -5,10 +5,8 @@ import {BeaconNode, BeaconDb, createNodeJsLibp2p} from "@lodestar/beacon-node";
 import {createIBeaconConfig} from "@lodestar/config";
 import {ACTIVE_PRESET, PresetName} from "@lodestar/params";
 import {IGlobalArgs} from "../../options/index.js";
-import {parseEnrArgs} from "../../options/enrOptions.js";
 import {onGracefulShutdown, getCliLogger} from "../../util/index.js";
-import {FileENR, overwriteEnrWithCliArgs, readPeerId} from "../../config/index.js";
-import {initializeOptionsAndConfig, persistOptionsAndConfig} from "../init/handler.js";
+import {initializeOptionsAndConfig} from "../init/handler.js";
 import {getVersionData} from "../../util/version.js";
 import {deleteOldPeerstorePreV036} from "../../migrations/index.js";
 import {IBeaconArgs} from "./options.js";
@@ -19,26 +17,10 @@ import {initBeaconState} from "./initBeaconState.js";
  * Runs a beacon node.
  */
 export async function beaconHandler(args: IBeaconArgs & IGlobalArgs): Promise<void> {
-  const {beaconNodeOptions, config} = await initializeOptionsAndConfig(args);
-  await persistOptionsAndConfig(args);
+  const {beaconNodeOptions, config, peerId} = await initializeOptionsAndConfig(args);
 
   const {version, commit} = getVersionData();
   const beaconPaths = getBeaconPaths(args);
-  // TODO: Rename db.name to db.path or db.location
-  beaconNodeOptions.set({db: {name: beaconPaths.dbDir}});
-  beaconNodeOptions.set({chain: {persistInvalidSszObjectsDir: beaconPaths.persistInvalidSszObjectsDir}});
-  // Add metrics metadata to show versioning + network info in Prometheus + Grafana
-  beaconNodeOptions.set({metrics: {metadata: {version, commit, network: args.network}}});
-  // Add detailed version string for API node/version endpoint
-  beaconNodeOptions.set({api: {version}});
-
-  // ENR setup
-  const peerId = await readPeerId(beaconPaths.peerIdFile);
-  const enr = FileENR.initFromFile(beaconPaths.enrFile, peerId);
-  const enrArgs = parseEnrArgs(args);
-  overwriteEnrWithCliArgs(enr, enrArgs, beaconNodeOptions.getWithDefaults());
-  const enrUpdate = !enrArgs.ip && !enrArgs.ip6;
-  beaconNodeOptions.set({network: {discv5: {enr, enrUpdate}}});
   const options = beaconNodeOptions.getWithDefaults();
 
   const abortController = new AbortController();
